fcm_backend/requirements.txt
firebase-admin>=6.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
pydantic-settings>=2.0.0
python-dotenv>=0.19.0
google-cloud-firestore>=2.3.0
google-auth>=2.3.0
requests>=2.26.0
pytest>=6.2.5
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.5
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
redis>=4.5.0
